Title: 10개 입력받기, Time: 98ms, Memory: 9MB, Status: Passed - Codetree

+++ 260210/10개 입력받기/receive-10-inputs.java
@@ -20,7 +20,7 @@ public class Main {
             count++;
         }
 
-        avg = sum/count;
+        avg = (double)sum/count;
 
         System.out.printf("%d %.1f", sum, avg);
